+++ R/getbb.R
@@ -141,27 +141,28 @@ bbox_to_string <- function (bbox) {
 #' @param key The API key to use for services that require it
 #' @param silent Should the API be printed to screen? TRUE by default
 #'
-#' @return Defaults to a matrix in the form:
-#' \code{
+#' @return For `format_out = "matrix"`, the default, return the bounding box:
+#' ```
 #'   min   max
 #' x ...   ...
 #' y ...   ...
-#' }
+#' ```
 #'
-#' If `format_out = "polygon"`, one or more two-columns matrices of polygonal
-#' longitude-latitude points. Where multiple `place_name` occurrences are found
-#' within `nominatim`, each item of the list of coordinates may itself contain
-#' multiple coordinate matrices where multiple exact matches exist. If one
-#' exact match exists with potentially multiple polygonal boundaries (for
-#' example, "london uk" is an exact match, but can mean either greater London or
-#' the City of London), only the first is returned. See examples below for
-#' illustration.
+#' If `format_out = "polygon"`, a list of polygons and multipolygons with one
+#' item for each `nominatim` result. The items are named with the OSM type and
+#' id. Each polygon is formed by one or more two-columns matrices of polygonal
+#' longitude-latitude points. The first matrix represents the outer boundary and
+#' the next ones represent holes. See examples below for illustration.
 #'
-#' For `format_out = "osm_type_id"`, a character string representing an OSM object in overpass query
-#' language. For example: \code{"relation(id:11747082)"} represents the area of
-#' the Catalan Countries. If one exact match exists with potentially multiple
-#' polygonal boundaries, only the first relation or way is returned. A set of
-#' objects can also be represented for multiple results (e.g.
+#' If `format_out = "sf_polygon"`, a `sf` object. Each row correspond to a
+#' `place_name` within `nominatim` result.
+#'
+#' For `format_out = "osm_type_id"`, a character string representing an OSM
+#' object in overpass query language. For example:
+#' `"relation(id:11747082)"` represents the area of the Catalan Countries.
+#' If one exact match exists with potentially multiple polygonal boundaries,
+#' only the first relation or way is returned. A set of objects can also be
+#' represented for multiple results (e.g.
 #' `relation(id:11747082,307833); way(id:22422490)`). See examples below for
 #' illustration. The OSM objects that can be used as
 #' [areas in overpass queries](https://wiki.openstreetmap.org/wiki/Overpass_API/Overpass_QL#Map_way/relation_to_area_(map_to_area))
@@ -187,14 +188,16 @@ bbox_to_string <- function (bbox) {
 #' getbb ("Hereford", format_out = "data.frame", limit = 3)
 #'
 #' # Examples of polygonal boundaries
-#' bb <- getbb ("london uk", format_out = "polygon") # single match
-#' dim (bb [[1]] [[1]]) # matrix of longitude/latitude pairs
+#' bb <- getbb ("Milano, Italy", format_out = "polygon")
+#' # A polygon and a multipolygon:
+#' str (bb) # matrices of longitude/latitude pairs
+#'
 #' bb_sf <- getbb ("kathmandu", format_out = "sf_polygon")
+#' bb_sf
 #' # sf:::plot.sf(bb_sf) # can be plotted if sf is installed
 #' getbb ("london", format_out = "sf_polygon")
-#' getbb ("accra", format_out = "sf_polygon") # rectangular bb
 #'
-#' area <- getbb ("València", format_out = "osm_type_id")
+#' getbb ("València", format_out = "osm_type_id")
 #' # select multiple areas with format_out = "osm_type_id"
 #' areas <- getbb ("València", format_out = "data.frame")
 #' bbox_to_string (areas [areas$osm_type != "node", ])
@@ -212,13 +215,17 @@ bbox_to_string <- function (bbox) {
 getbb <- function (place_name,
                    display_name_contains = NULL,
                    viewbox = NULL,
-                   format_out = "matrix",
+                   format_out = c (
+                       "matrix", "data.frame", "string",
+                       "polygon", "sf_polygon", "osm_type_id"
+                   ),
                    base_url = "https://nominatim.openstreetmap.org",
                    featuretype = "settlement",
                    limit = 10,
                    key = NULL,
                    silent = TRUE) {
 
+    format_out <- match.arg (format_out)
     is_polygon <- grepl ("polygon", format_out)
 
     obj <- get_nominatim_query (
@@ -261,38 +268,69 @@ getbb <- function (place_name,
         ret <- bb_mat
     } else if (format_out == "string") {
         ret <- bbox_to_string (bbox = bb_mat)
+
     } else if (is_polygon) {
 
         gt_p <- get_geotext_poly (obj)
         gt_mp <- get_geotext_multipoly (obj)
 
-        gt <- c (gt_p, gt_mp)
-        # multipolys below are not strict SF MULTIPOLYGONs, rather just cases
-        # where nominatim returns lists of multiple items
-        if (length (gt) == 0) {
-            message ("No polygonal boundary for ", place_name)
-            ret <- bb_mat
-        } else if (length (gt) == 1) {
-            ret <- gt [[1]]
-        } else {
-            ret <- gt
-        }
-    } else {
-        stop (paste0 (
-            "format_out not recognised; please specify one of ",
-            "[data.frame, matrix, string, polygon]"
-        ))
-    }
+        if (format_out == "polygon") {
 
-    if (format_out == "sf_polygon") {
-        ret_poly <- bb_as_sf_poly (gt_p, gt_mp, place_name)
-        obj_index <- as.integer (c (names (gt_p), names (gt_mp)))
-        ret_data <- obj [obj_index, which (!names (obj) %in% c ("boundingbox", "geotext"))]
-        ret <- cbind (ret_data, ret_poly)
-        # Then restore sf attributes:
-        nms <- names (ret)
-        attributes (ret) <- attributes (ret_poly)
-        names (ret) <- nms
+            gt <- c (gt_p, gt_mp)
+            # multipolys below are not strict SF MULTIPOLYGONs, rather just
+            # cases where nominatim returns lists of multiple items
+            if (length (gt) == 0) {
+                message (
+                    "No polygonal boundary for ", place_name,
+                    ". Returning the bounding box of the first result"
+                )
+                ret <- bb_mat
+            } else {
+                poly_id <- names (gt)
+                obj_id <- paste0 (obj$osm_type, "/", obj$osm_id)
+                obj_id <- intersect (obj_id, poly_id)
+                # sort geometries following Nominatim order
+                ord_poly <- match (obj_id, poly_id)
+                ret <- gt [ord_poly]
+            }
+        } else if (format_out == "sf_polygon") {
+
+            if (length (gt_p) + length (gt_mp) == 0) {
+                message (
+                    "No polygonal boundary for ", place_name,
+                    ". Returning the bounding boxes."
+                )
+                ret_poly <- lapply (obj$boundingbox, function (x) {
+                    x <- as.numeric (x)
+                    bb_mat <- matrix (
+                        c (x [3:4], x [1:2]),
+                        nrow = 2, byrow = TRUE
+                    )
+                    mat2sf_poly (list (bb_mat))
+                })
+                ret_poly <- do.call (rbind, ret_poly)
+                poly_id <- paste0 (obj$osm_type, "/", obj$osm_id)
+            } else {
+                ret_poly <- bb_as_sf_poly (gt_p, gt_mp)
+                poly_id <- c (names (gt_p), names (gt_mp))
+            }
+
+            obj_id <- paste0 (obj$osm_type, "/", obj$osm_id)
+            cols <- setdiff (names (obj), c ("boundingbox", "geotext"))
+            ret <- obj [obj_id %in% poly_id, cols]
+            obj_id <- intersect (obj_id, poly_id)
+
+            utf8cols <- c ("licence", "name", "display_name")
+            ret [, utf8cols] <- setenc_utf8 (ret [, utf8cols])
+
+            # sort geometries following Nominatim order
+            ord_poly <- match (obj_id, poly_id)
+            geometry <- ret_poly$geometry [ord_poly]
+            # sub-setting without 'sf' loaded removes attributes:
+            attributes (geometry) <- attributes (ret_poly$geometry)
+
+            ret <- make_sf (ret, geometry)
+        }
     }
 
     return (ret)
@@ -368,7 +406,8 @@ get_nominatim_query <- function (place_name,
 #' Get all polygons from a 'geojson' object
 #'
 #' @param obj A 'geojson' object
-#' @return List of polygon matrices
+#' @return List of polygons. Each polygon is a list of matrices, the first
+#'   defining the outer ring and the following ones, if present, define holes.
 #' @noRd
 get_geotext_poly <- function (obj) {
 
@@ -411,7 +450,22 @@ get_geotext_poly <- function (obj) {
         lens <- vapply (gt_p, length, integer (1))
         index_final <- rep (index_final, times = lens)
         gt_p <- do.call (c, gt_p)
-        names (gt_p) <- as.character (index_final)
+
+        # Aggregate rings by polygon
+        gt_p <- split (
+            gt_p,
+            paste0 (obj$osm_type [index_final], "/", obj$osm_id [index_final])
+        )
+        # Set names to the rings of the polygons
+        gt_p <- lapply (gt_p, function (x) {
+            if (length (x) > 1) {
+                inner <- paste0 ("inner_", seq_len (length (x) - 1))
+            } else {
+                inner <- character ()
+            }
+            names (x) <- c ("outer", inner)
+            x
+        })
     }
 
     return (gt_p)
@@ -422,7 +476,9 @@ get_geotext_poly <- function (obj) {
 #' See Issue #195
 #'
 #' @param obj A 'geojson' object
-#' @return List of multipolygon matrices
+#' @return List of multipolygons. Each multipolygon is a list of polygons, and
+#'   each polygon is a list of matrices where the first
+#'   defines the outer ring and the following ones, if present, define holes.
 #' @noRd
 get_geotext_multipoly <- function (obj) {
 
@@ -450,9 +506,27 @@ get_geotext_multipoly <- function (obj) {
 
     if (length (gt_mp) > 0) {
         gt_mp <- lapply (gt_mp, function (i) get1bdypoly (i))
-        lens <- vapply (gt_mp, length, integer (1))
-        index_final <- rep (index_final, times = lens)
-        names (gt_mp) <- as.character (index_final)
+
+        # Aggregate polygons by multipolygon
+        gt_mp <- split (
+            gt_mp,
+            paste0 (obj$osm_type [index_final], "/", obj$osm_id [index_final])
+        )
+        # Set names to the polygons and rings of the multypolygons
+        gt_mp <- lapply (gt_mp, function (x) {
+            names (x) <- paste0 ("pol_", seq_len (length (x)))
+            x <- lapply (x, function (y) {
+                if (length (y) > 1) {
+                    inner <- paste0 ("inner_", seq_len (length (y) - 1))
+                } else {
+                    inner <- character ()
+                }
+                names (y) <- c ("outer", inner)
+                y
+            })
+
+            x
+        })
     }
 
     return (gt_mp)
@@ -501,51 +575,53 @@ get1bdypoly <- function (p) {
     return (ret)
 }
 
-#' convert a matrix to an sf polygon
+#' Convert a list of matrices to an sf polygon
 #'
-#' @param mat A matrix
-#' @param pname The name of the polygon
+#' @param pol A list of matrices defining a polygon. First is the outer limit,
+#'   other are holes.
 #'
-#' @return A list that can be converted into a simple features geometry
+#' @return A `sf` object representing the polygon without using \pkg{sf}.
 #' @noRd
-mat2sf_poly <- function (mat, pname) {
-    if (nrow (mat) == 2) {
-        x <- c (mat [1, 1], mat [1, 2], mat [1, 2], mat [1, 1], mat [1, 1])
-        y <- c (mat [2, 2], mat [2, 2], mat [2, 1], mat [2, 1], mat [2, 2])
-        mat <- cbind (x, y)
+mat2sf_poly <- function (pol) {
+    if (length (pol) == 1L && nrow (pol [[1]]) == 2L) {
+        # no polygon but a bounding box
+        pol <- pol [[1]]
+        x <- c (pol [1, 1], pol [1, 2], pol [1, 2], pol [1, 1], pol [1, 1])
+        y <- c (pol [2, 2], pol [2, 2], pol [2, 1], pol [2, 1], pol [2, 2])
+        pol <- list (cbind (x, y))
     }
-    mat_sf <- list (mat)
-    class (mat_sf) <- c ("XY", "POLYGON", "sfg")
-    mat_sf <- list (mat_sf)
-    attr (mat_sf, "class") <- c ("sfc_POLYGON", "sfc")
-    attr (mat_sf, "precision") <- 0
-    bb <- as.vector (t (apply (mat, 2, range)))
+    class (pol) <- c ("XY", "POLYGON", "sfg")
+    pol_sf <- list (pol)
+    attr (pol_sf, "class") <- c ("sfc_POLYGON", "sfc")
+    attr (pol_sf, "precision") <- 0
+    bb <- as.vector (t (apply (do.call (rbind, pol), 2, range)))
     names (bb) <- c ("xmin", "ymin", "xmax", "ymax")
     class (bb) <- "bbox"
-    attr (mat_sf, "bbox") <- bb
+    attr (pol_sf, "bbox") <- bb
     crs <- list (
         input = "EPSG:4326",
         wkt = wkt4326
     )
     class (crs) <- "crs"
-    attr (mat_sf, "crs") <- crs
-    attr (mat_sf, "n_empty") <- 0L
-    mat_sf <- make_sf (mat_sf)
-    names (mat_sf) <- "geometry"
-    attr (mat_sf, "sf_column") <- "geometry"
-    return (mat_sf)
+    attr (pol_sf, "crs") <- crs
+    attr (pol_sf, "n_empty") <- 0L
+    pol_sf <- make_sf (pol_sf)
+    names (pol_sf) <- "geometry"
+    attr (pol_sf, "sf_column") <- "geometry"
+    return (pol_sf)
 }
 
 #' convert a list of matrices to an sf mulipolygon
 #'
 #' @param x A list of matrices
-#' @param mpname The name of the multipolygon
 #'
 #' @return A list that can be converted into a simple features geometry
 #' @noRd
-mat2sf_multipoly <- function (x, mpname) {
+mat2sf_multipoly <- function (x) {
     # get bbox from matrices
-    bb <- as.vector (t (apply (do.call (rbind, x [[1]]), 2, range)))
+    bb <- as.vector (t (apply (
+        do.call (rbind, unlist (x, recursive = FALSE)), 2, range
+    )))
     names (bb) <- c ("xmin", "ymin", "xmax", "ymax")
     class (bb) <- "bbox"
 
@@ -567,29 +643,29 @@ mat2sf_multipoly <- function (x, mpname) {
     return (xsf)
 }
 
-bb_as_sf_poly <- function (gt_p, gt_mp, place_name) {
+bb_as_sf_poly <- function (gt_p, gt_mp) {
 
     if (!is.null (gt_p)) {
         gt_p <- lapply (gt_p, function (i) {
-            mat2sf_poly (i, pname = place_name)
+            mat2sf_poly (i)
         })
     }
     if (!is.null (gt_mp)) {
         gt_mp <- lapply (gt_mp, function (i) {
-            mat2sf_multipoly (list (i), mpname = place_name)
+            mat2sf_multipoly (i)
         })
     }
 
-    if (is.null (gt_p) & is.null (gt_mp)) {
+    if (length (gt_p) == 0 && length (gt_mp) == 0) {
         stop ("Query returned no polygons")
-    } else if (is.null (gt_mp)) {
+    } else if (length (gt_mp) == 0) {
         ret <- do.call (rbind, gt_p)
-    } else if (is.null (gt_p)) {
+    } else if (length (gt_p) == 0) {
         ret <- do.call (rbind, gt_mp)
     } else {
-        ret <- list (
-            "polygon" = do.call (rbind, gt_p),
-            "multipolygon" = do.call (rbind, gt_mp)
+        ret <- do.call (
+            rbind,
+            c (polygon = gt_p, multipolygon = gt_mp)
         )
     }